move `TestAssistedCandidateGeneratorDifferentTokenizers` into a new testing file
